refactor: 배너 제거 기능 추가 및 검색바에서 호출로 변경
