+++ src/view/moreButton.ts
@@ -10,22 +10,20 @@ const handleMoreButtonClick = async (page: number, total_pages: number, keyword?
   const skeletonList = renderSkeletons({ height: 300 });
   container.appendChild(skeletonList);
 
-  skeletonList.remove();
-
   const movies = await addMovies(page, keyword);
   container.appendChild(movies);
 
-  if (page >= total_pages) {
-    document.getElementById('moreButton')?.remove();
+  skeletonList.remove();
+
+  if (page === total_pages) {
+    $('#moreButton')?.remove();
   }
 };
 
 export const moreButton = (page: number, total_pages: number, keyword?: string) => {
-  const button = Button({
+  return Button({
     text: '더보기',
     id: 'moreButton',
     onClick: () => handleMoreButtonClick(++page, total_pages, keyword)
   });
-
-  return button;
 };
